// src/commands/entities/enums.ts
export enum OSType {
  UBUNTU = 'ubuntu',
  MACOS = 'macos',
  WINDOWS = 'windows',
}

export enum CommandType {
  PS = 'ps',
  TASKLIST = 'tasklist',
}

export enum ProcessStatus {
  RUNNING = 'running',
  SLEEPING = 'sleeping',
  STOPPED = 'stopped',
  ZOMBIE = 'zombie',
  IDLE = 'idle',
  UNKNOWN = 'unknown',
}